# src/main/resources/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: MM-CQ-Java
  description: Modular monolith with Command & Query
  version: "0.1"
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearer: []
paths:
  /api/customer/login:
    post:
      tags:
      - Mobile API
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/com.mgleska.mmcqjava2.customer.action.command.LoginCmd.ParamDto"
        required: true
      responses:
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.shared.CustomExceptionHandler.ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.customer.action.command.LoginCmd.ResultDto"
      security: []
  /api/customer/change-store:
    post:
      tags:
      - Mobile API
      operationId: changeStore
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/com.mgleska.mmcqjava2.customer.action.command.ChangeStoreCmd.ParamDto"
        required: true
      responses:
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.shared.CustomExceptionHandler.ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.customer.action.command.ChangeStoreCmd.ResultDto"
  /api/admin/user/login:
    post:
      tags:
      - Admin
      operationId: login_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/com.mgleska.mmcqjava2.user.action.command.AdminLoginCmd.ParamDto"
        required: true
      responses:
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.shared.CustomExceptionHandler.ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.user.action.command.AdminLoginCmd.ResultDto"
      security: []
  /api/admin/offer/change-visibility:
    post:
      tags:
      - Admin
      operationId: changeVisibility
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/com.mgleska.mmcqjava2.offer.action.command.ChangeVisibilityCmd.ParamDto"
        required: true
      responses:
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.shared.CustomExceptionHandler.ErrorResponse"
        "200":
          description: OK
  /api/store/{id}:
    get:
      tags:
      - Mobile API
      operationId: details
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.shared.CustomExceptionHandler.ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.store.action.query.GetStoreDetailsQry.ResultDto"
  /api/store/list:
    get:
      tags:
      - Mobile API
      operationId: list
      responses:
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.shared.CustomExceptionHandler.ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/com.mgleska.mmcqjava2.store.action.query.GetStoreListQry.ListItemDto"
  /api/offer/{id}:
    get:
      tags:
      - Mobile API
      operationId: details_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.shared.CustomExceptionHandler.ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.offer.action.query.GetOfferDetailsQry.ResultDto"
  /api/offer/list:
    get:
      tags:
      - Mobile API
      operationId: list_1
      parameters:
      - name: queryParam
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/com.mgleska.mmcqjava2.offer.action.query.GetOfferListQry.ParamDto"
      responses:
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.shared.CustomExceptionHandler.ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.offer.action.query.GetOfferListQry.ResultDto"
  /api/admin/offer/{id}:
    get:
      tags:
      - Admin
      operationId: details_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.shared.CustomExceptionHandler.ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.offer.action.query.AdminGetOfferDetailsQry.ResultDto"
  /api/admin/offer/list:
    get:
      tags:
      - Admin
      operationId: list_2
      parameters:
      - name: queryParam
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/com.mgleska.mmcqjava2.offer.action.query.AdminGetOfferListQry.ParamDto"
      responses:
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.shared.CustomExceptionHandler.ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.offer.action.query.AdminGetOfferListQry.ResultDto"
  /api/admin/offer/filters:
    get:
      tags:
      - Admin
      operationId: filters
      responses:
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.shared.CustomExceptionHandler.ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/com.mgleska.mmcqjava2.offer.action.query.AdminGetListFiltersQry.ResultDto"
components:
  schemas:
    com.mgleska.mmcqjava2.shared.CustomExceptionHandler.ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        error:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        errors:
          type: array
          items:
            $ref: "#/components/schemas/com.mgleska.mmcqjava2.shared.CustomExceptionHandler.FieldError"
        path:
          type: string
          minLength: 1
        timestamp:
          type: string
          minLength: 1
      required:
      - error
      - message
      - path
      - status
      - timestamp
    com.mgleska.mmcqjava2.shared.CustomExceptionHandler.FieldError:
      type: object
      properties:
        field:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
      required:
      - field
      - message
    com.mgleska.mmcqjava2.customer.action.command.LoginCmd.ParamDto:
      type: object
      properties:
        customerId:
          type: integer
          format: int32
      required:
      - customerId
    com.mgleska.mmcqjava2.customer.action.command.LoginCmd.ResultDto:
      type: object
      properties:
        token:
          type: string
          minLength: 1
      required:
      - token
    com.mgleska.mmcqjava2.customer.action.command.ChangeStoreCmd.ParamDto:
      type: object
      properties:
        storeId:
          type: integer
          format: int32
      required:
      - storeId
    com.mgleska.mmcqjava2.customer.action.command.ChangeStoreCmd.ResultDto:
      type: object
      properties:
        token:
          type: string
          minLength: 1
      required:
      - token
    com.mgleska.mmcqjava2.user.action.command.AdminLoginCmd.ParamDto:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
      required:
      - email
    com.mgleska.mmcqjava2.user.action.command.AdminLoginCmd.ResultDto:
      type: object
      properties:
        token:
          type: string
          minLength: 1
      required:
      - token
    com.mgleska.mmcqjava2.offer.action.command.ChangeVisibilityCmd.ParamDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        version:
          type: integer
          format: int32
        visible:
          type: boolean
      required:
      - id
      - version
      - visible
    com.mgleska.mmcqjava2.store.action.query.GetStoreDetailsQry.ResultDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        externalId:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
      required:
      - address
      - externalId
      - id
      - name
    com.mgleska.mmcqjava2.store.action.query.GetStoreListQry.ListItemDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        externalId:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
      required:
      - externalId
      - id
      - name
    com.mgleska.mmcqjava2.offer.action.query.GetOfferDetailsQry.ResultDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        productEan:
          type: string
          minLength: 1
        productName:
          type: string
          minLength: 1
        price:
          type: integer
          format: int32
        lowestPrice:
          type: integer
          format: int32
        imageUrl:
          type: string
        quantityLevel:
          type: string
          enum:
          - UNKNOWN
          - AVAILABLE
          - AVAILABLE_LOW
      required:
      - id
      - imageUrl
      - price
      - productEan
      - productName
      - quantityLevel
    com.mgleska.mmcqjava2.offer.action.query.GetOfferListQry.ParamDto:
      type: object
      properties:
        page:
          type: integer
          format: int32
    com.mgleska.mmcqjava2.offer.action.query.GetOfferListQry.ResultDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/com.mgleska.mmcqjava2.offer.action.query.GetOfferListQry.ResultItemDto"
        page:
          type: integer
          format: int32
        perPage:
          type: integer
          format: int32
      required:
      - items
      - page
      - perPage
    com.mgleska.mmcqjava2.offer.action.query.GetOfferListQry.ResultItemDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        productEan:
          type: string
          minLength: 1
        productName:
          type: string
          minLength: 1
        price:
          type: integer
          format: int32
        lowestPrice:
          type: integer
          format: int32
      required:
      - id
      - price
      - productEan
      - productName
    com.mgleska.mmcqjava2.offer.action.query.AdminGetOfferDetailsQry.ResultDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        version:
          type: integer
          format: int32
        visible:
          type: boolean
        productEan:
          type: string
          minLength: 1
        productName:
          type: string
          minLength: 1
        price:
          type: integer
          format: int32
        lowestPrice:
          type: integer
          format: int32
        imageUrl:
          type: string
          minLength: 1
        quantity:
          type: integer
          format: int32
      required:
      - id
      - imageUrl
      - price
      - productEan
      - productName
      - version
      - visible
    com.mgleska.mmcqjava2.offer.action.query.AdminGetOfferListQry.ParamDto:
      type: object
      properties:
        search:
          type: string
        storeId:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
        perPage:
          type: integer
          format: int32
    com.mgleska.mmcqjava2.offer.action.query.AdminGetOfferListQry.ResultDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/com.mgleska.mmcqjava2.offer.action.query.AdminGetOfferListQry.ResultItemDto"
        page:
          type: integer
          format: int32
        perPage:
          type: integer
          format: int32
      required:
      - items
      - page
      - perPage
    com.mgleska.mmcqjava2.offer.action.query.AdminGetOfferListQry.ResultItemDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        storeName:
          type: string
          minLength: 1
        visible:
          type: boolean
        productEan:
          type: string
          minLength: 1
        productName:
          type: string
          minLength: 1
        price:
          type: integer
          format: int32
      required:
      - id
      - price
      - productEan
      - productName
      - storeName
      - visible
    com.mgleska.mmcqjava2.offer.action.query.AdminGetListFiltersQry.ResultDto:
      type: object
      properties:
        stores:
          type: array
          items:
            $ref: "#/components/schemas/com.mgleska.mmcqjava2.offer.action.query.AdminGetListFiltersQry.ResultItemDto"
      required:
      - stores
    com.mgleska.mmcqjava2.offer.action.query.AdminGetListFiltersQry.ResultItemDto:
      type: object
      properties:
        storeId:
          type: integer
          format: int32
        storeName:
          type: string
          minLength: 1
      required:
      - storeId
      - storeName
  securitySchemes:
    bearer:
      type: http
      name: bearer
      scheme: bearer
      bearerFormat: JWT